Move config hash/validation to pallet itself
